refactor(backend): pg-mem testing helper 분리
